refactor(core): migrate JIT to use deps tracker behind a flag
The current change is done behind a flag which is set to false. So no change in code path took place here. In a followup PR the flag is changed to true which will make the actual change.
