+++ packages/core/src/render3/deps_tracker/api.ts
@@ -6,7 +6,7 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import {Type} from '../../core';
+import {Type} from '../../interface/type';
 import {NgModuleType} from '../../metadata/ng_module_def';
 import {ComponentType, DependencyTypeList, DirectiveType, NgModuleScopeInfoFromDecorator, PipeType} from '../interfaces/definition';
 
@@ -84,7 +84,7 @@ export interface DepsTrackerApi {
    * The main application of this method is for test beds where we want to clear the cache to
    * enforce scope update after overriding.
    */
-  clearScopeCacheFor(type: ComponentType<any>|NgModuleType): void;
+  clearScopeCacheFor(type: Type<any>): void;
 
   /**
    * Returns the scope of NgModule. Mainly to be used by JIT and test bed.
